+++ broker/hosts.py
@@ -1,9 +1,14 @@
 # from functools import cached_property
-from broker.settings import settings
+from logzero import logger
 from broker import session
+from broker.exceptions import NotImplementedError
+from broker.settings import settings
 
 
 class Host:
+
+    default_timeout = 0  # timeout in ms, 0 is infinite
+
     def __init__(self, hostname, name=None, from_dict=False, **kwargs):
         # Allow the class to construct itself from kwargs
         if from_dict:
@@ -40,8 +45,12 @@ class Host:
                 return mgr
         return None
 
-    def execute(self, command):
-        return self.session.run(command)
+    def execute(self, command, timeout=None):
+        timeout = timeout or self.default_timeout
+        logger.debug(f"{self.hostname} executing command: {command}")
+        res = self.session.run(command, timeout=timeout)
+        logger.debug(f"{self.hostname} command result:\n{res}")
+        return res
 
     def to_dict(self):
         return {
@@ -51,6 +60,14 @@ class Host:
             "type": "host",
             "_broker_args": self._broker_args,
         }
+
+    def setup(self):
+        """Automatically ran when entering a VMBroker context manager"""
+        pass
+
+    def teardown(self):
+        """Automatically ran when exiting a VMBroker context manager"""
+        pass
 
     @classmethod
     def from_dict(cls, arg_dict):